WIP: Almost done.
This commit implements some of the missing todo's that still had to be done.
It is printing out all the nodes, but I am not sure if it is right. 3 stays the root
the whole time and that doesn't seem right. I am going to do it by hand, but it is close.